requirements.txt
requests==2.22.0
simplejson==3.6.4
bs4==0.0.1
beautifulsoup4==4.3.2
pika==1.1.0
python-dotenv==0.10.3
numpy==1.16.4
opencv-python
pymongo==3.4.0